fix: fixing navigation paths, relocating components
