+++ src/pages/Auctions/AuctionCancelConfirmationPage.tsx
@@ -2,7 +2,7 @@ import { arrowBackOutline } from 'ionicons/icons';
 import { FC } from 'react';
 import { useHistory, useParams } from 'react-router';
 import styled from 'styled-components';
-import Button from '../components/ionic/Button';
+import Button from '../../components/ionic/Button';
 import {
   IonCol,
   IonContent,
@@ -15,11 +15,11 @@ import {
   IonFooter,
   IonRouterLink
 } from '@ionic/react';
-import Header from '../components/Header';
-import Text from '../components/ionic/Text';
-import { FontSize } from '../constants/font-size';
-import { FontWeight } from '../constants/font-weight';
-import { driverHighResImage } from '../constants/images';
+import Header from '../../components/Header';
+import Text from '../../components/ionic/Text';
+import { FontSize } from '../../constants/font-size';
+import { FontWeight } from '../../constants/font-weight';
+import { driverHighResImage } from '../../constants/images';
 
 const ImageBgCol = styled(IonCol)`
   display: flex;
