+++ src/pages/Auctions/AuctionMyAuctionExpiredAndAcceptOfferViewPage.tsx
@@ -5,10 +5,10 @@ import { useForm } from 'react-hook-form';
 import { shallowEqual, useSelector } from 'react-redux';
 import { useHistory, useParams } from 'react-router';
 import styled from 'styled-components';
-import { FontSize } from '../constants/font-size';
-import { FontWeight } from '../constants/font-weight';
-import Title from '../components/ionic/Title';
-import Button from '../components/ionic/Button';
+import { FontSize } from '../../constants/font-size';
+import { FontWeight } from '../../constants/font-weight';
+import Title from '../../components/ionic/Title';
+import Button from '../../components/ionic/Button';
 import { JSX } from '@ionic/core';
 import {
   IonButton,
@@ -23,10 +23,10 @@ import {
   IonSpinner,  
 } from '@ionic/react';
 import { BaseResourcesTypes } from '@protokol/client';
-import Header from '../components/Header';
-import Img from '../components/ionic/Img';
-import { addIPFSGatewayPrefix } from '../constants/images';
-import { collectionSelector } from '../store/selectors/collections';
+import Header from '../../components/Header';
+import Img from '../../components/ionic/Img';
+import { addIPFSGatewayPrefix } from '../../constants/images';
+import { collectionSelector } from '../../store/selectors/collections';
 
 const Footer = styled(IonFooter)`
   position: fixed;
