+++ src/pages/Auctions/AuctionMyAuctionViewPage.tsx
@@ -5,11 +5,11 @@ import { Controller, useForm } from 'react-hook-form';
 import { shallowEqual, useSelector } from 'react-redux';
 import { useHistory, useParams } from 'react-router';
 import styled from 'styled-components';
-import { Styles } from '../utils/styles';
-import { FontSize } from '../constants/font-size';
-import { FontWeight } from '../constants/font-weight';
-import Title from '../components/ionic/Title';
-import Button from '../components/ionic/Button';
+import { Styles } from '../../utils/styles';
+import { FontSize } from '../../constants/font-size';
+import { FontWeight } from '../../constants/font-weight';
+import Title from '../../components/ionic/Title';
+import Button from '../../components/ionic/Button';
 import { JSX } from '@ionic/core';
 
 import {
@@ -29,10 +29,10 @@ import {
   IonSpinner,  
 } from '@ionic/react';
 import { BaseResourcesTypes } from '@protokol/client';
-import Header from '../components/Header';
-import Img from '../components/ionic/Img';
-import { addIPFSGatewayPrefix } from '../constants/images';
-import { collectionSelector } from '../store/selectors/collections';
+import Header from '../../components/Header';
+import Img from '../../components/ionic/Img';
+import { addIPFSGatewayPrefix } from '../../constants/images';
+import { collectionSelector } from '../../store/selectors/collections';
 
 const Footer = styled(IonFooter)`
   position: fixed;
